made color codes to be used as floats

+++ app/config/prod.py
@@ -7,9 +7,9 @@ ENV_FILE_PATH = BASE_DIR / ".env"
 
 load_dotenv(ENV_FILE_PATH)
 
-RED = int(os.environ.get("RED", 1))
-GREEN = int(os.environ.get("GREEN", 1))
-BLUE = int(os.environ.get("BLUE", 0))
+RED = float(os.environ.get("RED", 1))
+GREEN = float(os.environ.get("GREEN", 1))
+BLUE = float(os.environ.get("BLUE", 0))
 
 class ProdConfig:
     RED = RED
